requirements.txt
opencv-python
Pillow
face-recognition
